ICU-20000 Changing BigInteger to BigDecimal

+++ icu4j/main/classes/core/src/com/ibm/icu/number/Scale.java
@@ -42,7 +42,7 @@ public class Scale {
             // Attempt to convert the BigDecimal to a magnitude multiplier.
             // ICU-20000: JDKs have inconsistent behavior on stripTrailingZeros() for Zero.
             arbitrary =
-                arbitrary.compareTo(BigInteger.ZERO) == 0
+                arbitrary.compareTo(BigDecimal.ZERO) == 0
                     ? BigDecimal.ZERO
                     : arbitrary.stripTrailingZeros();
             if (arbitrary.precision() == 1 && arbitrary.unscaledValue().equals(BigInteger.ONE)) {
